mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
